The Great Refactor
- Refactored some pesky functions outside the Gallery component; should be more compact and easy to adjust now
- Added basic unit tests
- Added GalleryContext to reduce prop drilling; some work should still need to be done though
