+++ package.json
@@ -27,7 +27,7 @@
     "deploy": "gh-pages -d build",
     "start": "vite",
     "build": "vite build",
-    "test": "react-scripts test",
+    "test": "vitest",
     "eject": "react-scripts eject",
     "dev": "vite"
   },
@@ -55,8 +55,10 @@
     "@svgr/plugin-jsx": "^8.1.0",
     "@vitejs/plugin-react": "^4.6.0",
     "gh-pages": "^6.2.0",
+    "jsdom": "^29.1.1",
     "tailwindcss": "^4.1.11",
     "vite": "^7.1.2",
+    "vitest": "^4.1.9",
     "wrangler": "^4.56.0"
   }
 }
