+++ src/api/galleryAPI.jsx
@@ -1,8 +1,11 @@
 
 import { useQuery } from '@tanstack/react-query'
 import { convertTagToSelectLabel, handleFilterParams } from "../stylesfunctions/apiFilter"
+import { useSearchParams } from 'react-router-dom'
+import handleUrlQuery from '../stylesfunctions/handleUrlQuery'
+import { useGallery } from '../components/galleryContext'
 
-export function portfolioData(){
+export function portfolioData() {
     const result = useQuery({
         queryKey: ["portfolioData"],
         queryFn: async () => {
@@ -11,12 +14,11 @@ export function portfolioData(){
         }
 
     })
-
     return result
 
 }
 
-export function portfolioTagData(){
+export function portfolioTagData() {
     const result = useQuery({
         queryKey: ["portfolioTagData"],
         queryFn: async () => {
@@ -27,7 +29,7 @@ export function portfolioTagData(){
     return result
 }
 
-export function groupTagData(){
+export function groupTagData() {
     const result = useQuery({
         queryKey: ["groupTagData"],
         queryFn: async () => {
@@ -36,4 +38,81 @@ export function groupTagData(){
         }
     })
     return result
+}
+
+export function handlePageViaSearchParams() {
+    const searchParams = new URLSearchParams(window.location.search)
+
+    let paramPage = searchParams.get("page")
+    if (paramPage === null || isNaN(paramPage)) {
+        paramPage = 1
+
+        searchParams.append("page", 1)
+    }
+}
+
+//todo: move handlefilterparams helper to be same place as here, or vice versa?
+export function handlePageFilters(portfolioTagData, searchParams = new URLSearchParams(window.location.search)) {
+    let tempSelectedFilters = ["include", "exclude"]
+        .reduce((accumulator, filterType) => {
+            return (handleFilterParams(filterType, accumulator, portfolioTagData, searchParams))
+        }, {
+            include: [],
+            exclude: []
+        })
+
+
+    new Array("include", "exclude").forEach((filterType) => {
+        if (tempSelectedFilters[filterType] === null) {
+            return
+        }
+        if (tempSelectedFilters[filterType].length === 0) {
+            searchParams.delete(filterType)
+        }
+    })
+
+    return tempSelectedFilters
+}
+
+export function writeTagsToParams(selectedFilters) {
+    const searchParams = new URLSearchParams(window.location.search)
+
+    let tempSelectedFilters = { ...selectedFilters }
+    let includeFilters = tempSelectedFilters.include
+    let excludeFilters = tempSelectedFilters.exclude
+
+
+    function handleFilter(filter) {
+        let tempValue = filter.value
+        tempValue = tempValue.split(" ").join("+")
+        return tempValue
+    }
+
+    includeFilters = includeFilters.map((filter) => {
+        return handleFilter(filter)
+    })
+
+    excludeFilters = excludeFilters.map((filter) => {
+        return handleFilter(filter)
+    })
+
+    includeFilters = includeFilters.join(",")
+    excludeFilters = excludeFilters.join(",")
+
+    let urlString = window.location.search
+    urlString = handleUrlQuery(urlString, "include", includeFilters)
+    urlString = handleUrlQuery(urlString, "exclude", excludeFilters)
+
+
+    console.log(urlString)
+    history.pushState({}, "Gallery", urlString)
+
+    /*
+    if (includeFilters !== "") {
+        searchParams.append("include", includeFilters)
+    }
+    if (excludeFilters !== "") {
+        searchParams.append("exclude", excludeFilters)
+    }*/
+
 }
