+++ src/stylesfunctions/apiFilter.js
@@ -6,24 +6,23 @@ export const convertTagToSelectLabel = (tagName, portfolioData) => {
     }
 }
 
-export const handleFilterParams = (filterType, tempSearchParams, selectedFilters, portfolioTagData) => {
-    let paramFilter = tempSearchParams.get(filterType)
+export const handleFilterParams = (filterType, selectedFilters, portfolioTagData, searchParams) => {
+    let paramFilter = searchParams.get(filterType)
     let tempSelectedFilters = selectedFilters
     if (paramFilter !== null) {
         paramFilter = paramFilter.split(",")
-
         tempSelectedFilters[filterType] = paramFilter.map((filterName) => {
             try {
                 return convertTagToSelectLabel(filterName, portfolioTagData)
             } catch (error) {
+                console.log(filterName)
                 console.log("dame da error")
                 console.log(error)
 
             }
-        console.log(tempSelectedFilters)
         }).filter((filterName) => {
             return filterName !== undefined
         })
     }
-    return tempSelectedFilters 
+    return tempSelectedFilters
 }
